Some documentation changes and finishing up config_manager
