+++ src/args.rs
@@ -6,25 +6,24 @@ use clap::{command, Parser};
     version, 
     about, 
     long_about = r###"
-Queries can be piped in through stdin and will be added to the context
-first. 
+cgip is a command-line tool that leverages OpenAI's models to address and 
+respond to user queries. It compiles context for these queries by prioritizing 
+input in a specific order: stdin, command-line arguments, and then files which 
+require you using the -f flag. 
 
-The query passed in as the first argument will then be added to the context
-second.
-
-Files read using -f will be added last to the context.
+The config file will be located in your config directory under cgip/config.toml.
+Please edit this file or use `cgip config --set key=value` to set your configuration.
 
 Usage examples:
 - To pipe input in from ls, run `ls | cgip "what can you tell me about these files?"`
   This will pipe the output of ls in as the first bit of context and then add the user
   query to the context.
 
-More detailed information can go here.
 "###
 )]
 pub struct Args {
-    /// What you want to ask Chat GPT. Query is optional but is added to the prompt
-    /// after the contents of the file or stdin if present.
+    /// Optional. The primary query to sent to the model. 
+    /// This is added to the context after stdin and file input.
     #[arg(index=1)]
     pub query: Option<String>,
 
@@ -33,20 +32,20 @@ pub struct Args {
     #[arg(short, long)]
     pub file: Option<String>,
 
-    /// The model to use. Defaults to `gpt-4`
+    /// Specify model to use. Defaults to `gpt-4`.
     #[arg(short = 'M', long)]
     pub model: Option<String>,
     
-    /// List the available models
+    /// List all the available models. 
     #[arg(short, long)]
     pub list_models:bool,
 
-    /// Show progress indicator (might mess up stdout)
+    /// Show progress indicator (might mess up stdout).
     #[arg(short = 'p', long)]
     pub show_progress: bool,
     
-    /// Prints not only the output from OpenAI but the chat context with all
-    /// assistant and user messages.
+    /// Output the full context used in the query, including chat context with
+    /// all assistant and user messages.
     #[arg(short = 'c',long)]
     pub show_context: bool,
     
@@ -62,7 +61,7 @@ pub struct Args {
 pub enum SubCommands {
     /// Render the context without running a query against the model.
     View(ViewSubCommand),
-    /// Set or get configuration values.
+    /// Set or get default configuration values with your config.toml.
     Config(ConfigSubCommand),
 }
 
@@ -75,11 +74,13 @@ pub struct ViewSubCommand {
 #[derive(Parser, Debug)]
 #[command(author, version, about = "Set or get configuration values", long_about = None)]
 pub struct ConfigSubCommand {
-    /// Set a configuration value in your config.toml 
+    /// Set a configuration value. Use the format key=value.
+    /// `cgip config --set model=gpt-4-turbo`
     #[arg(short, long)]
     pub set: Option<String>,
 
-    /// Get your current configuration value
+    /// Get your current configuration value. 
+    /// `cgip config --get model`
     #[arg(short, long)]
     pub get: Option<String>,
 }
